+++ utils/CartFetcher.ts
@@ -8,17 +8,21 @@ export class CartFetcher extends BaseCartFetcher {
   static async fetchCart(request: Request, actionContext: ActionContext): Promise<Cart> {
     const cartApi = new CartApi(actionContext.frontasticContext, getLocale(request));
 
-    if (request.sessionData?.account !== undefined) {
-      return (await cartApi.getForUser(request.sessionData.account, request.sessionData.organization) as Cart);
-    }
-
     if (request.sessionData?.cartId !== undefined) {
       try {
-        return (await cartApi.getById(request.sessionData.cartId) as Cart);
+        const cart = (await cartApi.getById(request.sessionData.cartId) as Cart);
+        if (cartApi.assertCartOrganization(cart, request.sessionData.organization)) {
+          return cart;
+        }
       } catch (error) {
         console.info(`Error fetching the cart ${request.sessionData.cartId}, creating a new one. ${error}`);
       }
     }
+    
+    if (request.sessionData?.account !== undefined) {
+      return (await cartApi.getForUser(request.sessionData.account, request.sessionData.organization) as Cart);
+    }
+
     // @ts-ignore
     return {};
   }